\name{Bross-package}
\alias{Bross-package}
\alias{Bross}
\docType{package}
\title{
\packageTitle{Bross}
}
\description{
\packageDescription{Bross}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Bross}
\packageIndices{Bross}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Bross}

Maintainer: \packageMaintainer{Bross}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
