% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster}
\alias{cluster}
\title{Cluster Analysis}
\usage{
cluster(data, using_distance, using_method, n_cluster)
}
\arguments{
\item{data}{Un dataframe non vuoto}

\item{using_distance}{Un metodo valido per il calcolo della distanza}

\item{using_method}{Un metodo valido per la clusterizzazione dei dati}

\item{n_cluster}{Un numero positivo per la creazione dei cluster}
}
\value{
Stampa del dendogramma

Una lista con i dati originali con aggiunta una colonna di classe ($data)

La numerosita' dei singoli cluster ($group)

I baricentri dei singoli cluster (dati standardizzati) ($baricentri_cluster)

Informazioni sulla compattezza dei singoli cluster ($compattezza_cluster)

Informazioni relative all'inerzia dei cluster (interna, totale e fra classi) ($inerzia)
}
\description{
Prende un dataframe e ne esegue la clusterizzazione in base alle direttive date
}
\examples{
## Eseguiamo una cluster analysis delle unita' sperimentali del dataset USArrests
## e salviamo i risultati in una variabile. Impostiamo come distanza quella euclidea,
## utilizziamo il metodo di Ward e vogliamo creare 4 cluster

result <- cluster(datasets::USArrests, "euclidean", "ward.D", 4)

## Verrà stampata la numerosita' dei singoli gruppi. Per richiamare i risultati basta scrivere:

result$data
result$group
result$baricentri_cluster
result$compattezza_cluster
result$inerzia
}
