% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_cor.R
\name{cov_cor}
\alias{cov_cor}
\title{Covarianza e Correlazione variabili quantitative}
\usage{
cov_cor(data)
}
\arguments{
\item{data}{Un dataset di sole variabili quantitative (di tipo numeric)}
}
\value{
Tabella covarianza di tutte le variabili

Tabella correlazione di tutte le variabili
}
\description{
Preso un dataframe di variabili di solo tipo numeric ne calcola le tabelle di covarianza e correlazione
}
\examples{
## Calcoliamo la covarianza e correlazione tra le variabili del dataset USArrests
## e le salviamo in una variabile

result <- cov_cor(datasets::USArrests)

## Verranno stampati i risultati. Per richiamarli basta scrivere:

result$COV_TOT
result$COR_TOT
}
