% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bross_check.R
\name{divide_FACTOR.NUMERIC}
\alias{divide_FACTOR.NUMERIC}
\title{Divisione variabili factor e numeric}
\usage{
divide_FACTOR.NUMERIC(data)
}
\arguments{
\item{data}{Un dataframe non vuoto}
}
\value{
Un oggetto contenente due dataframe: uno di sole variabili di tipo factor e uno di sole numeric
}
\description{
Preso un dataframe, viene restituito un oggetto di due componenti avente un dataframe di sole variabili factor ed uno di sole variabili numeric
}
\examples{
## Dividiamo il dataset esoph in variabili di tipo factor e di tipo numeric
## Salviamo in una variabile i risultati ottenuti

result <- divide_FACTOR.NUMERIC(datasets::esoph)

## Richiamiamo i risultati ottenuti:

result$factor
result$numeric
}
