% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indici_conf.R
\name{indici_conf}
\alias{indici_conf}
\title{Indici di Confronto variabili quantitative}
\usage{
indici_conf(data)
}
\arguments{
\item{data}{Un dataset di sole variabili quantitative (di tipo numeric)}
}
\value{
Tabella indici di confronto della media/mediana e quartili di tutte le variabili
}
\description{
Preso un dataframe di variabili di solo tipo numeric ne calcola gli indici di confronto
}
\examples{
## Calcoliamo gli indici di confronto delle variabili del dataset USArrests
## e li salviamo in una variabile

result <- indici_conf(datasets::USArrests)

## Verranno stampati i risultati. Per richiamarli basta scrivere:

result
}
