% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot}
\alias{pca_plot}
\title{PCA Plot}
\usage{
pca_plot(data)
}
\arguments{
\item{data}{Un dataframe non vuoto}
}
\value{
Stampa dei singoli grafici delle coppie di componenti

Le tabelle di fedelta' dei singoli grafici
}
\description{
Prende un dataframe e ne esegue la PCA per poi stampare i grafici delle componenti piu' rilevanti
}
\examples{
## Stampiamo i plot della PCA delle sole componenti ritenute le piu' significative,
## ovvero sulle quali sono meglio rappresentate le variabili (si sceglie il minor numero
## possibile), lavorando sul dataset USArrests.

pca_plot(datasets::USArrests)

## Verrà stampata la tabella di fedelta' della PCA, i plot riguardo le componenti d'interesse
## e le relative tabelle di fedelta' di rappresentazione dei singoli grafici
}
\seealso{
\code{\link{pca}} Metodo utilizzato per l'analisi in componenti principali

\code{\link{pca_plot_comp}} Metodo utilizzato per la stampa dei grafici delle componenti e la creazione delle tabelle di fedelta' dei grafici
}
