% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_plot.R
\name{pca_plot_comp}
\alias{pca_plot_comp}
\title{PCA Plot N-Componenti}
\usage{
pca_plot_comp(data, R, n_comp)
}
\arguments{
\item{data}{Un dataframe non vuoto}

\item{R}{Una matrice R corretta per il dataframe dato}

\item{n_comp}{Numero di componenti da prendere in considerazione}
}
\value{
Stampa dei singoli grafici delle coppie di componenti

Le tabelle di fedelta' dei singoli grafici
}
\description{
Prende un dataframe e ne esegue la PCA per poi stampare i grafici delle componenti scelte e le tabelle di fedelta' dei grafici
}
\examples{
## Stampiamo i plot della PCA delle prime n_comp componenti selezionate, lavorando sul dataset
## USArrests e prendendo la matrice R calcolata con la funzione pca()

R <- pca(datasets::USArrests)$R_PCA
pca_plot_comp(datasets::USArrests, R, 2)

## Verrà stampata la tabella di fedelta' della PCA, i plot riguardo le componenti d'interesse
## e le relative tabelle di fedelta' di rappresentazione dei singoli grafici
}
