% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantili.R
\name{quantili}
\alias{quantili}
\title{Quantili variabili quantitative}
\usage{
quantili(data)
}
\arguments{
\item{data}{Un dataset di sole variabili quantitative (di tipo numeric)}
}
\value{
Tabella dei quantili di tutte le variabili
#' @examples
## Calcoliamo i quantili di tutte le variabili del dataset USArrests
## e li salviamo in una variabile

result <- quantili(datasets::USArrests)

## Verranno stampati i risultati. Per richiamarli basta scrivere:

result
}
\description{
Preso un dataframe di variabili di solo tipo numeric ne calcola i quantili principali (MIN, Q1, Q2, Q3, MAX)
}
