% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_CV.R
\name{sd_cv}
\alias{sd_cv}
\title{Standard Deviation e Coefficiente di Variazione variabili quantitative}
\usage{
sd_cv(data)
}
\arguments{
\item{data}{Un dataset di sole variabili quantitative (di tipo numeric)}
}
\value{
Tabella Standard Deviation e Coefficiente di Variazione di tutte le variabili
}
\description{
Preso un dataframe di variabili di solo tipo numeric ne calcola la Standard Deviation e il Coefficiente di Variazione
}
\examples{
## Calcoliamo la standard deviation e il coefficiente di variazioni delle variabili del dataset
## USArrests e li salviamo in una variabile

result <- sd_cv(datasets::USArrests)

## Verranno stampati i risultati. Per richiamarli basta scrivere:

result
}
