% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_factor.R
\name{table_factor}
\alias{table_factor}
\title{Analisi di due variabili factor}
\usage{
table_factor(factor1, factor2, name, position)
}
\arguments{
\item{factor1}{prima variabile di tipo factor}

\item{factor2}{seconda variabile di tipo factor}

\item{name}{un vettore di due posizioni contenente i nomi con i quali verranno rappresentati le due variabili factor}

\item{position}{la posizione della legenda nella rappresentazione dei barplot}
}
\value{
tutte le informazioni relative ai due fattori

la stampa dei barplot delle condizionate
}
\description{
Prende due variabili factor e ne calcola le tabelle assolute e relative prende in considerazione i marginali,
calcola le condizionate riga e colonna e ne disegna i barplot, infine calcola le deviazioni riga e colonna
}
\examples{
## Prendiamo due variabili qualitative di tipo fattore, un vettore di due posizioni contenente
## i nomi di tali variabili (utile per la stampa) e la posizione della legenda nel grafico
## (in alto a destra). Utilizziamo le variabili "agegp" (Gruppo di eta') e "alcgp"
## (Quantitativo di consumo di alcool) provenienti dal dataset esoph

dataset <- datasets::esoph
table_factor(dataset$agegp, dataset$alcgp, c("Fascia_eta", "Alcool"), "topright")

## Vengono stampate le tabelle relative per i due fattori separati, le tabelle assolute e relative
## con i marginali, profili riga e colonna e le deviazioni riga e colonna. Vengono inoltre stampati
## i barplot relativi alle condizionate riga e colonna
}
